consolidate jobs to single job class for refactoring
